move classes to util package
